requirements.txt
python-telegram-bot[job-queue]==20.8
google-generativeai==0.8.3
supabase==2.10.0
gtts==2.5.1
python-dotenv==1.0.1
pytz==2025.2
edge-tts==7.2.3
Flask==3.0.0
